move to RecipeCompatUtil
